https://www.tinkercad.com/things/7vYCex558R5-amazing-allis-uusam/editel
The smart dustbin is a carefully designed solution that solves the social issue of waste disposal, the smart dustbin identifies the kind of material being thrown inside it and segregates it into bio or non biodegradable. The dustbin also comes with an option to provide wifi as an incentive of throwing garbage.